+++ src/handless/exceptions.py
@@ -1,3 +1,6 @@
+from typing import Any
+
+
 class HandlessException(Exception):  # noqa: N818
     """Base exception for all handless errors."""
 
@@ -5,8 +8,15 @@ class HandlessException(Exception):  # noqa: N818
 class BindingNotFoundError(HandlessException):
     """When no binding is registered for a given type."""
 
-    def __init__(self, type_: type) -> None:
+    def __init__(self, type_: type[Any]) -> None:
         super().__init__(f"There is no binding registered for {type_}")
+
+
+class BindingAlreadyExistingError(HandlessException):
+    """When trying to register an already registered type."""
+
+    def __init__(self, type_: type[Any]) -> None:
+        super().__init__(f"There is already a binding registered for {type_}")
 
 
 class ResolveError(HandlessException):